Change money props to decimal from int

+++ CMP 205 Unit 1 WPF Application for Comic Book Store scenario/DundeeComicBookStore/Interfaces/ProductInterface.cs
@@ -8,8 +8,8 @@ namespace DundeeComicBookStore.Interfaces
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public uint UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public uint UnitsInStock { get; set; }
-        public uint UnitCost { get; set; }
+        public decimal UnitCost { get; set; }
     }
 }
