+++ CMP 205 Unit 1 WPF Application for Comic Book Store scenario/DundeeComicBookStore/Models/ProductModel.cs
@@ -9,8 +9,8 @@ namespace DundeeComicBookStore.Models
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public uint UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public uint UnitsInStock { get; set; }
-        public uint UnitCost { get; set; }
+        public decimal UnitCost { get; set; }
     }
 }
